+++ changelog.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
 
+- MarkdownEncoder does not trim spaces for Pre tags
 - Func GenerateID and TOC related funcs generate uniq ids for similar titles
 
 ## [0.20.0] 2021-07-25